Create branch, change MultiPin Playback function, create README
Previously, MultiPin_Playback created major memory overflow errors whenever more than ~10 notes existed in a song. This was due to the 8kB usable RAM on the micro:bit, which couldn't contain the 100n, 24kB per note array created by the Create_Millisecond_Music function.

Obviously the function was extremely inefficient, so instead the Play_Dual_Soundfunction was changed to simply play one given note n times. Create_Millisecond_Music was scrapped in favor of Create_Single_Step_List, which just expands the given list so each note is only length 1. This, too, originally proved to overload the memory, so the function also deletes the original list as it goes, to save on memory.

The program also suffered from playback timing issues, likely due to the BPM of 600, and Tick length of 100. The music module was probably never made with that in mind, so it didn't seem to work well. I simply took it down an order of magnitude, and shortened the play duration.